Fix ty failures on Python 3.10 surfaced by the new CI matrix leg
numpy's typing stubs resolve np.logspace's return dtype less precisely
on the numpy version that installs for Python 3.10 (floating[Any])
than on 3.12 (float64), which ty then rejects against the strict
NDArray[np.float64] lambda_grid parameters in _core.pyi. This was
already latent in pipeline.py's default Subtractor.lambda_grid and in
benchmarks/diag_real_data.py -- invisible until this PR's CI matrix
change actually ran ty against 3.10 for the first time.

Pin dtype=np.float64 explicitly at each np.logspace/np.asarray call
that feeds a float64-typed core function (pipeline.py's default
lambda_grid, plus the affected tests and benchmark script). Also fix
an unrelated, equally latent ty error in catalog.py: np.empty(n,
dtype=CATALOG_DTYPE) isn't recognised as a structured (void-dtype)
array since CATALOG_DTYPE is a runtime np.dtype value rather than a
literal, so the by-name field assignment failed to type-check; cast
it explicitly.

+++ python/delta/pipeline.py
@@ -324,7 +324,9 @@ class Subtractor:
         self.saturation = saturation
         self.bright_mask = bright_mask
         self.lambda_grid = (
-            np.logspace(-6.0, 6.0, 25) if lambda_grid is None else np.asarray(lambda_grid)
+            np.logspace(-6.0, 6.0, 25, dtype=np.float64)
+            if lambda_grid is None
+            else np.asarray(lambda_grid, dtype=np.float64)
         )
         self.clip_sigma = clip_sigma
         self.clip_iterations = clip_iterations
